get rid of unneeded inbox views for now
