+++ src/cljs/athens/router.cljs
@@ -156,7 +156,6 @@
 (def routes
   ["/"
    ["" {:name :home}]
-   #_["notifications" {:name :notifications}]
    ["settings" {:name :settings}]
    ["pages" {:name :pages}]
    ["page-t/:title" {:name :page-by-title}]
