--- src/cljs/athens/views/notifications.cljs
@@ -1,8 +1,0 @@
-#_(ns athens.views.notifications
-    (:require
-      ["/components/Inbox/Inbox" :refer [Inbox]]))
-
-
-#_(defn inbox
-    []
-    [:> Inbox])
