+++ src/cljs/athens/views/pages/core.cljs
@@ -3,7 +3,6 @@
     ["@chakra-ui/react" :refer [Box]]
     [athens.util :refer [toast]]
     [athens.views.hoc.perf-mon :as perf-mon]
-    #_[athens.views.notifications :as notifications]
     [athens.views.pages.all-pages :as all-pages]
     [athens.views.pages.daily-notes :as daily-notes]
     [athens.views.pages.graph :as graph]
@@ -79,7 +78,5 @@
                        [daily-notes/page]]
        :graph         [perf-mon/hoc-perfmon-no-new-tx {:span-name "pages/graph"}
                        [graph/page]]
-       #_#_:notifications         [perf-mon/hoc-perfmon-no-new-tx {:span-name "pages/notifications"}
-                                   [notifications/inbox]]
        [perf-mon/hoc-perfmon-no-new-tx {:span-name "pages/default"}
         [daily-notes/page]])]))
